# Use `values` iterator  and `update` in map equality/hash