--- a/pkgs/collection/CHANGELOG.md
+++ b/pkgs/collection/CHANGELOG.md
@@ -2,10 +2,8 @@
 
 - Add `IterableMapEntryExtension` for working on `Map` as a list of pairs, using
   `Map.entries`.
-- Use `entries` and `update` in map equality implementations.
-  - Speeds up all equality checks.
-  - Speeds up implementations for maps with slow lookup, at a slight cost for
-    other implementations.
+- Optimize equality and hash code for maps by using `update` and a `values`
+  iterator to avoid extra lookups.
 
 ## 1.19.1