--- a/pkgs/collection/lib/src/equality.dart
+++ b/pkgs/collection/lib/src/equality.dart
@@ -325,16 +325,15 @@ class MapEquality<K, V> implements Equality<Map<K, V>> {
     var length = map1.length;
     if (length != map2.length) return false;
     Map<_MapEntry, int> equalElementCounts = HashMap();
-    for (var key in map1.keys) {
-      var entry = _MapEntry(this, key, map1[key]);
-      var count = equalElementCounts[entry] ?? 0;
-      equalElementCounts[entry] = count + 1;
+    for (var MapEntry(:key, :value) in map1.entries) {
+      var entry = _MapEntry(this, key, value);
+      equalElementCounts.update(entry, (i) => i + 1, ifAbsent: () => 1);
     }
-    for (var key in map2.keys) {
-      var entry = _MapEntry(this, key, map2[key]);
-      var count = equalElementCounts[entry];
-      if (count == null || count == 0) return false;
-      equalElementCounts[entry] = count - 1;
+    for (var MapEntry(:key, :value) in map2.entries) {
+      var entry = _MapEntry(this, key, value);
+      var count =
+          equalElementCounts.update(entry, (i) => i - 1, ifAbsent: () => -1);
+      if (count < 0) return false;
     }
     return true;
   }
@@ -343,9 +342,9 @@ class MapEquality<K, V> implements Equality<Map<K, V>> {
   int hash(Map<K, V>? map) {
     if (map == null) return null.hashCode;
     var hash = 0;
-    for (var key in map.keys) {
+    for (var MapEntry(:key, :value) in map.entries) {
       var keyHash = _keyEquality.hash(key);
-      var valueHash = _valueEquality.hash(map[key] as V);
+      var valueHash = _valueEquality.hash(value);
       hash = (hash + 3 * keyHash + 7 * valueHash) & _hashMask;
     }
     hash = (hash + (hash << 3)) & _hashMask;