@jakemac53 2024-11-05
If we don't want to land the `entries` changes, I will end up having to fork this implementation. But, that is probably OK, just a bit unfortunate.